refactor: 과제별 패키지 분리
