+++ src/main/java/com/sryoondev/racingcar/step2/MyNumber.java
@@ -1,4 +1,4 @@
-package com.sryoondev.racingcar;
+package com.sryoondev.racingcar.step2;
 
 import java.util.Objects;
