+++ src/main/java/com/sryoondev/racingcar/step2/MyOperator.java
@@ -1,4 +1,4 @@
-package com.sryoondev.racingcar;
+package com.sryoondev.racingcar.step2;
 
 public class MyOperator {
     String operator;
