+++ src/main/java/com/sryoondev/racingcar/step3/Route.java
@@ -1,5 +1,4 @@
-package com.sryoondev.racingcar;
-
+package com.sryoondev.racingcar.step3;
 
 import java.util.ArrayList;
